pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mealmetric"
version = "0.1.0"
description = "MealMetric backend MVP scaffolding"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi>=0.110.0,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.0,<3.0.0",
  "psycopg2-binary>=2.9.0,<3.0.0",
  "alembic>=1.13.0,<2.0.0",
  "pydantic-settings>=2.2.0,<3.0.0",
  "email-validator>=2.0.0,<3.0.0",
  "prometheus-client>=0.20.0,<1.0.0",
  "python-dotenv>=1.0.0,<2.0.0",
  "bcrypt>=4.0.0,<6.0.0",
  "stripe>=8.0.0,<9.0.0",
]

[project.optional-dependencies]
dev = [
  "black>=24.0.0,<26.0.0",
  "ruff>=0.6.0,<1.0.0",
  "mypy>=1.10.0,<2.0.0",
  "types-psycopg2>=2.9.21.2024.1,<3.0.0",
  "pytest>=8.0.0,<9.0.0",
  "pytest-cov>=5.0.0,<7.0.0",
  "httpx>=0.27.0,<1.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py312"]
extend-exclude = "(alembic|tools)"

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]
extend-exclude = ["alembic", "tools"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "C4", "SIM", "N", "ASYNC"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["mealmetric"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
disallow_any_generics = true
no_implicit_optional = true
show_error_codes = true
pretty = true
exclude = ["^alembic/"]

[[tool.mypy.overrides]]
module = ["prometheus_client.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q --cov=mealmetric --cov-report=term-missing --cov-fail-under=90"
testpaths = ["tests"]
